flask_face_api/requirements.txt
flask
flask-restx
flask-cors
flask-pydantic
pymongo
pymysql
mysql-connector-python
bcrypt
pyjwt
pydantic
loguru
python-dotenv
gunicorn
fpdf2
pandas
matplotlib
seaborn
numpy